cli/requirements.txt
requests
pyyaml
gitpython
boto3
tqdm
jsonschema
botocore
python-dateutil
paramiko
git-url-parse
rich
rich-argparse
fastapi
starlette
uvicorn
pydantic
sqlalchemy[asyncio]
py-cpuinfo
psutil
websocket-client
cursor
simple-term-menu
jinja2
pygtail
google-cloud-storage
google-cloud-compute
google-cloud-secret-manager
google-cloud-logging
aiosqlite
packaging
alembic>=1.10.2
apscheduler
typing-extensions
file-read-backwards
psutil
azure-identity>=1.12.0
azure-keyvault-secrets>=4.6.0
azure-storage-blob>=12.15.0
azure-mgmt-subscription>=3.1.1
azure-mgmt-compute>=29.1.0
azure-mgmt-network==23.0.0b2
azure-mgmt-resource>=22.0.0
azure-mgmt-authorization>=3.0.0
azure-mgmt-storage>=21.0.0
azure-mgmt-keyvault>=10.1.0
azure-mgmt-loganalytics
azure-mgmt-msi
azure-mgmt-monitor
azure-graphrbac